requirements.txt
telethon==1.42.0
faster-whisper==1.1.1
anthropic>=0.40.0
fpdf2>=2.8.1
python-dotenv>=1.0.0
pyyaml>=6.0.0
tqdm>=4.66.0

# Web UI
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
jinja2>=3.1.0
python-multipart>=0.0.9
sse-starlette>=2.0.0
